+++ inst/case_studies/mixture_iris.R
@@ -4,9 +4,7 @@ library(PLNmodels)
 library(tidyverse)
 library(viridisLite)
 
-nb_cores <- 10
-
-## Dimension reduction with PCA
+nb_cores <- 4
 
 count <- iris %>% dplyr::select(-Species) %>% exp() %>% round() %>% as.matrix()
 covariate <- data.frame(Species = iris$Species)
@@ -16,8 +14,8 @@ my_mixtures <-  PLNmixture(Abundance ~ 1 + offset(log(Offset)), clusters = 1:5, 
 plot(my_mixtures)
 
 myPLN <- getBestModel(my_mixtures)
-myPLN$plot_clustering_pca()
-myPLN$plot_clustering_data()
+plot(myPLN, type = "pca")
+plot(myPLN, type = "matrix")
 
 aricode::ARI(myPLN$memberships, iris$Species)
 
@@ -26,5 +24,5 @@ my_mixtures_covar <-  PLNmixture(Abundance ~ 0 + Species + offset(log(Offset)), 
 plot(my_mixtures_covar)
 
 myPLN_covar <- getBestModel(my_mixtures_covar)
-myPLN_covar$plot_clustering_pca()
-myPLN_covar$plot_clustering_data()
+plot(myPLN_covar, "pca")
+plot(myPLN_covar, "matrix")